In Description
***SERVER***
Added endpoints implemented in NoificationOperations:
* GET - notifications;
* GET - notifications/unreadcount
* PUT - notifications/:id

Added controls that check if moderator is validated in UserOperations and InsertionOperations

***CLIENT***
edit-insertion.component.ts: Added code to check if insertion is != null and it is not closed after it is get from the server and if the current user is the mod (validated) or the insertionist

edit-profile.component.ts: Added method in ngOnInit that redirects to page not found if there is not a token

insertion-component.html: Now the message.post.component and the reply button of every message are visible if there is a token

insertion.component.ts: Added code that checks if insertion is == null and if it is, it redirects to page-not-found.component

login.component.ts:  Added line of code that tells the navbar to reload the notificationscount after the login has been done. The communication is done with notification-http.service

message-reply-component.ts: redirects to page-not-found.component.ts if there is not a token or if the insertion that has been get is null or if there isn't the message in :m_id in that insertion

navbar-component.html: added buttons to go to the new components

navbar-component.ts: added public variable notificationsCount and methods to get the notification count.
In ngOninit added subscription to observable that tells to call notificationscount when something happens

post-insertion.component: reroutes to pagenotfound.component if there is no token or the token has a moderator role set as true

signupmod.component: reroutes to pagenotfoundcomponent if the user is not a mod and it is not validated

userlist.component: reroutes to pagenotfound.component if there is no token or the user is not a validated moderator

notification-http.service: ADDED